refactored timer comp for reusability
